# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BSV DEX Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades/history:
    get:
      operationId: getTradeHistory
      tags: [trades]
      summary: Get user trade history
      parameters:
        - in: query
          name: walletAddress
          required: true
          schema:
            type: string
        - in: query
          name: symbol
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Get user portfolio/balances
      parameters:
        - in: query
          name: walletAddress
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Portfolio data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
...
  /wallet/connect:
    post:
      operationId: connectWallet
      tags: [wallet]
      summary: Register wallet connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectWalletRequest"
      responses:
        "200":
          description: Wallet connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletSession"
  /wallet/transactions:
    get:
      operationId: getWalletTransactions
      tags: [wallet]
      summary: Get wallet on-chain transactions
      parameters:
        - in: query
          name: walletAddress
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Wallet transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletTransaction"
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        price:
          type: number
        quantity:
          type: number
        total:
          type: number
        fee:
          type: number
        feeAsset:
          type: string
        timestamp:
          type: string
          format: date-time
        txid:
          type: string
          description: BSV on-chain transaction ID
        walletAddress:
          type: string
      required:
        - id
        - symbol
        - side
        - price
        - quantity
        - total
        - fee
        - feeAsset
        - timestamp
...
    Portfolio:
      type: object
      properties:
        walletAddress:
          type: string
        totalValueUSD:
          type: number
        totalPnlUSD:
          type: number
        totalPnlPercent:
          type: number
        balances:
          type: array
          items:
            $ref: "#/components/schemas/AssetBalance"
        openOrdersCount:
          type: integer
        openPositionsCount:
          type: integer
      required:
        - walletAddress
        - totalValueUSD
        - totalPnlUSD
        - totalPnlPercent
        - balances
        - openOrdersCount
        - openPositionsCount
    AssetBalance:
      type: object
      properties:
        asset:
          type: string
        free:
          type: number
        locked:
          type: number
        total:
          type: number
        valueUSD:
          type: number
        pnl24h:
          type: number
        pnl24hPercent:
          type: number
      required:
        - asset
        - free
        - locked
        - total
        - valueUSD
...
    ConnectWalletRequest:
      type: object
      properties:
        address:
          type: string
          description: BSV wallet address
        publicKey:
          type: string
          description: BSV public key
        provider:
          type: string
          enum: [handcash, relayx, twetch, panda, yours, sensilet]
        signature:
          type: string
          description: Signed message to verify wallet ownership
        message:
          type: string
          description: The message that was signed
      required:
        - address
        - provider
    WalletSession:
      type: object
      properties:
        address:
          type: string
        provider:
          type: string
        connected:
          type: boolean
        connectedAt:
          type: string
          format: date-time
        publicKey:
          type: string
      required:
        - address
        - provider
        - connected
        - connectedAt
    WalletTransaction:
      type: object
      properties:
        txid:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, trade, fee, funding]
        asset:
          type: string
        amount:
          type: number
        fee:
          type: number
        status:
          type: string
          enum: [confirmed, pending, failed]
        confirmations:
          type: integer
        timestamp:
          type: string
          format: date-time
        blockHeight:
          type: integer
      required:
        - txid
        - type
        - asset
        - amount
        - status
        - timestamp