# lib/api-zod/src/generated/types/connectWalletRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BSV DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectWalletRequestProvider } from "./connectWalletRequestProvider";

export interface ConnectWalletRequest {
  /** BSV wallet address */
  address: string;
  /** BSV public key */
  publicKey?: string;
  provider: ConnectWalletRequestProvider;
  /** Signed message to verify wallet ownership */
  signature?: string;
  /** The message that was signed */
  message?: string;
}