# lib/api-zod/src/generated/types/walletTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BSV DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletTransactionStatus } from "./walletTransactionStatus";
import type { WalletTransactionType } from "./walletTransactionType";

export interface WalletTransaction {
  txid: string;
  type: WalletTransactionType;
  asset: string;
  amount: number;
  fee?: number;
  status: WalletTransactionStatus;
  confirmations?: number;
  timestamp: Date;
  blockHeight?: number;
}